refactor: 예약 내역 페이지 리팩토링
